feat: implement generic YAML configuration loader and update trigger loaders
